chore(deps): upgrade tree-sitter to 0.26, rusqlite to 0.40
tree-sitter 0.20 → 0.26.9 (ABI 15, supports 13–15):
- Grammar crates dropped language() free-fn; now export LANGUAGE: LanguageFn.
  chunk_with_parser now takes impl Into<Language> and converts at call site.
- node.child() now takes u32 (was usize); both traversal loops updated.
- Grammar versions: rust 0.24.2, python/go/js 0.25.0, cpp 0.23.4
  (cpp uses tree-sitter-language 0.1 as runtime dep — compatible with 0.26 core)

rusqlite 0.31 → 0.40.0:
- usize no longer implements ToSql; cast limit to i64 in search_fts params.
- cc bumped to 1.2.63 automatically (no ort conflict: ort is optional/default-off).

136/136 tests pass; clippy clean; release build succeeds.

+++ src/store.rs
@@ -738,7 +738,7 @@ pub fn search_fts(
             "SELECT c.id FROM chunks c JOIN chunks_fts f ON c.id = f.rowid
              WHERE chunks_fts MATCH ?1 AND instr(c.path, ?2) > 0 LIMIT ?3",
         )?;
-        let rows = stmt.query_map(params![sanitized, filter, limit], |row| {
+        let rows = stmt.query_map(params![sanitized, filter, limit as i64], |row| {
             row.get::<_, i64>(0)
         })?;
         for id in rows.flatten() {
@@ -747,7 +747,7 @@ pub fn search_fts(
     } else {
         let mut stmt =
             conn.prepare("SELECT rowid FROM chunks_fts WHERE chunks_fts MATCH ?1 LIMIT ?2")?;
-        let rows = stmt.query_map(params![sanitized, limit], |row| row.get::<_, i64>(0))?;
+        let rows = stmt.query_map(params![sanitized, limit as i64], |row| row.get::<_, i64>(0))?;
         for id in rows.flatten() {
             ids.push(id);
         }
